+++ client/src/components/DataFiles/DataFilesModals/DataFilesModals.js
@@ -8,6 +8,8 @@ import DataFilesRenameModal from './DataFilesRenameModal';
 import DataFilesPushKeysModal from './DataFilesPushKeysModal';
 import DataFilesCopyModal from './DataFilesCopyModal';
 import DataFilesTrashModal from './DataFilesTrashModal';
+import DataFilesLinkModal from './DataFilesLinkModal';
+import DataFilesShowPathModal from './DataFilesShowPathModal';
 import './DataFilesModals.scss';
 
 export default function DataFilesModals() {
@@ -21,6 +23,8 @@ export default function DataFilesModals() {
       <DataFilesRenameModal />
       <DataFilesPushKeysModal />
       <DataFilesTrashModal />
+      <DataFilesLinkModal />
+      <DataFilesShowPathModal />
       <SystemsPushKeysModal />
     </>
   );