+++ client/src/redux/reducers/systemMonitor.reducers.js
@@ -1,4 +1,4 @@
-const initialState = {
+export const initialState = {
   list: [],
   loading: false,
   error: null