+++ server/portal/apps/datafiles/views.py
@@ -1,14 +1,24 @@
-from portal.apps.auth.tasks import get_user_storage_systems
-from portal.views.base import BaseApiView
+import json
+import logging
 from django.conf import settings
 from django.http import JsonResponse, HttpResponseForbidden
 from requests.exceptions import HTTPError
-import json
-import logging
+from portal.apps.auth.tasks import get_user_storage_systems
+from portal.views.base import BaseApiView
 from portal.apps.datafiles.handlers.tapis_handlers import (tapis_get_handler,
                                                            tapis_put_handler,
                                                            tapis_post_handler)
 from portal.apps.users.utils import get_allocations
+from portal.apps.accounts.managers.user_systems import UserSystemsManager
+from portal.apps.datafiles.models import Link
+from portal.exceptions.api import ApiException
+from django.contrib.auth.decorators import login_required
+from django.utils.decorators import method_decorator
+from portal.libs.agave.models.systems.storage import StorageSystem
+from portal.libs.agave.serializers import BaseAgaveSystemSerializer
+from .utils import notify, NOTIFY_ACTIONS
+import logging
+
 
 logger = logging.getLogger(__name__)
 
@@ -21,14 +31,13 @@ class SystemListingView(BaseApiView):
         local_systems = settings.PORTAL_DATA_DEPOT_LOCAL_STORAGE_SYSTEMS
 
         user_systems = get_user_storage_systems(request.user.username, local_systems)
-
         # compare available storage systems to the systems a user can access
         response = {'system_list': []}
-        for _, details in user_systems.items():
+        for system_name, details in user_systems.items():
             response['system_list'].append(
                 {
                     'name': details['name'],
-                    'system': details['prefix'].format(request.user.username),
+                    'system':  UserSystemsManager(request.user, system_name=system_name).get_system_id(),
                     'scheme': 'private',
                     'api': 'tapis',
                     'icon': details['icon']
@@ -38,7 +47,16 @@ class SystemListingView(BaseApiView):
         default_system = user_systems[settings.PORTAL_DATA_DEPOT_LOCAL_STORAGE_SYSTEM_DEFAULT]
         response['default_host'] = default_system['host']
 
-        return JsonResponse(response)
+        for system in response['system_list']:
+            try:
+                if system['api'] == 'tapis' and 'system' in system:
+                    system['definition'] = StorageSystem(
+                        request.user.agave_oauth.client, id=system['system']
+                    )
+            except Exception:
+                logger.exception("Could not retrieve definition for {}".format(system['system']))
+
+        return JsonResponse(response, encoder=BaseAgaveSystemSerializer)
 
 
 class TapisFilesView(BaseApiView):
@@ -50,9 +68,13 @@ class TapisFilesView(BaseApiView):
         try:
             response = tapis_get_handler(
                 client, scheme, system, path, operation, **request.GET.dict())
+
+            operation in NOTIFY_ACTIONS and \
+                notify(request.user.username, operation, 'success', {'response': response})
         except HTTPError as e:
             error_status = e.response.status_code
             error_json = e.response.json()
+            operation in NOTIFY_ACTIONS and notify(request.user.username, operation, 'error', {})
             if error_status == 502:
                 # In case of 502 determine cause
                 system = dict(client.systems.get(systemId=system))
@@ -72,14 +94,19 @@ class TapisFilesView(BaseApiView):
 
     def put(self, request, operation=None, scheme=None,
             handler=None, system=None, path='/'):
-
         body = json.loads(request.body)
         try:
             client = request.user.agave_oauth.client
         except AttributeError:
             return HttpResponseForbidden
 
-        response = tapis_put_handler(client, scheme, system, path, operation, body=body)
+        try:
+            response = tapis_put_handler(client, scheme, system, path, operation, body=body)
+            operation in NOTIFY_ACTIONS and \
+                notify(request.user.username, operation, 'success', {'response': response})
+        except Exception as exc:
+            operation in NOTIFY_ACTIONS and notify(request.user.username, operation, 'error', {})
+            raise exc
 
         return JsonResponse({"data": response})
 
@@ -91,6 +118,82 @@ class TapisFilesView(BaseApiView):
         except AttributeError:
             return HttpResponseForbidden()
 
-        response = tapis_post_handler(client, scheme, system, path, operation, body=body)
+        try:
+            response = tapis_post_handler(client, scheme, system, path, operation, body=body)
+            operation in NOTIFY_ACTIONS and \
+                notify(request.user.username, operation, 'success', {'response': response})
+        except Exception as exc:
+            operation in NOTIFY_ACTIONS and notify(request.user.username, operation, 'error', {})
+            raise exc
 
+        return JsonResponse({"data": response})
+
+
+@method_decorator(login_required, name='dispatch')
+class LinkView(BaseApiView):
+    def create_postit(self, request, scheme, system, path):
+        client = request.user.agave_oauth.client
+        body = {
+            "url": "{tenant}/files/v2/media/system/{system}/{path}".format(
+                tenant=settings.AGAVE_TENANT_BASEURL,
+                system=system,
+                path=path
+            ),
+            "unlimited": True
+        }
+        response = client.postits.create(body=body)
+        postit = response['_links']['self']['href']
+        link = Link.objects.create(
+            agave_uri=f"{system}/{path}",
+            postit_url=postit
+        )
+        link.save()
+        return postit
+
+    def delete_link(self, request, link):
+        client = request.user.agave_oauth.client
+        response = client.postits.delete(uuid=link.get_uuid())
+        link.delete()
+        return response
+
+    def get(self, request, scheme, system, path):
+        """Given a file, returns a link for a file
+        """
+        try:
+            link = Link.objects.get(agave_uri=f"{system}/{path}")
+        except Link.DoesNotExist:
+            return JsonResponse({"data": None})
+        return JsonResponse({"data": link.postit_url})
+
+    def delete(self, request, scheme, system, path):
+        """Delete an existing link for a file
+        """
+        try:
+            link = Link.objects.get(agave_uri=f"{system}/{path}")
+        except Link.DoesNotExist:
+            raise ApiException("Post-it does not exist")
+        response = self.delete_link(request, link)
+        return JsonResponse({"data": response})
+
+    def post(self, request, scheme, system, path):
+        """Generates a new link for a file
+        """
+        try:
+            Link.objects.get(agave_uri=f"{system}/{path}")
+        except Link.DoesNotExist:
+            # Link doesn't exist - proceed with creating one
+            response = self.create_postit(request, scheme, system, path)
+            return JsonResponse({"data": response})
+        # Link for this file already exists, raise an exception
+        raise ApiException("Link for this file already exists")
+
+    def put(self, request, scheme, system, path):
+        """Replace an existing link for a file
+        """
+        try:
+            link = Link.objects.get(agave_uri=f"{system}/{path}")
+            self.delete_link(request, link)
+        except Link.DoesNotExist:
+            raise ApiException("Could not find pre-existing link")
+        response = self.create_postit(request, scheme, system, path)
         return JsonResponse({"data": response})