+++ server/portal/apps/onboarding/api/urls.py
@@ -2,15 +2,13 @@
 .. :module:: apps.accounts.api.urls
    :synopsis: Manager handling anything pertaining to accounts
 """
-from django.conf.urls import url
-from portal.apps.onboarding.api.views import (
-    SetupStepView,
-    SetupAdminView
-)
+from django.urls import path
+from portal.apps.onboarding.api import views
 
 
 app_name = 'portal_onboarding_api'
 urlpatterns = [
-    url(r'^user/(?P<username>[\w]+)/?$', SetupStepView.as_view(), name='user_view'),
-    url(r'^admin/$', SetupAdminView.as_view(), name='user_admin')
+    path('user/', views.SetupStepView.as_view(), name='user_self_view'),
+    path('user/<str:username>/', views.SetupStepView.as_view(), name='user_view'),
+    path('admin/', views.SetupAdminView.as_view(), name='user_admin')
 ]